feat - add posts controller
