+++ .gitignore
@@ -4,6 +4,7 @@
 dist
 tmp
 /out-tsc
+.env
 
 # dependencies
 node_modules
